MAINT: Reorganize the builtin datasource types and standardizes the file-based datasource types (CSV, Dataset, Zarr, and H5PY)
