+++ podpac/core/data/test/test_dataset.py
@@ -1,0 +1,11 @@
+import os.path
+import numpy as np
+
+from podpac.core.data.file import Dataset
+
+
+class TestDataset(object):
+    """ test xarray Dataset source
+    """
+
+    pass
